moved home components to portolio dir
